requirements.txt
alembic==1.13.3
aniso8601==9.0.1
apispec==6.6.1
attrs==24.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.0.4
